[silmtensor] rename slimtensor fiies to consistent with executorch files
Pull Request resolved: https://github.com/pytorch/executorch/pull/16469

This diff updates the slimtensor file names to follow ET style (use under_score instead of Camel).
Note that we keep the c10 files as Camel since et/c10 files are also in Camel Case.
ghstack-source-id: 336658382
@exported-using-ghexport

Differential Revision: [D90143740](https://our.internmc.facebook.com/intern/diff/D90143740/)
