+++ backends/aoti/tests/test_common_shims_slim.cpp
@@ -12,7 +12,7 @@
 #include <executorch/backends/aoti/common_shims_slim.h>
 #include <executorch/backends/aoti/slim/c10/core/Device.h>
 #include <executorch/backends/aoti/slim/c10/core/ScalarType.h>
-#include <executorch/backends/aoti/slim/factory/Empty.h>
+#include <executorch/backends/aoti/slim/factory/empty.h>
 #include <executorch/runtime/core/error.h>
 #include <executorch/runtime/platform/platform.h>
